+++ examples/cpp/qst/qst_example.cpp
@@ -24,6 +24,7 @@ int main() {
         sim
     );
     QuantumStateTomography<SimpleCircuitExecution> qstworkflow(workflow); //wrap into QST object
+    qstworkflow.set_maximum_likelihood_estimation(); //enable maximum likelihood estimation
 
     //(3) pass to metric
     QuantumStateDensity<QuantumStateTomography<SimpleCircuitExecution>> metric(qstworkflow);